// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_Task = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    description: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    priority: 'low' | 'medium' | 'high';
    status: 'todo' | 'in_progress' | 'done';
    title: CodecTypes['pg/text@1']['output'];
    updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    userId: CodecTypes['pg/int4@1']['output'];
    user: public_User;
    readonly [RelationKeys]?: 'user';
  };
  export type public_User = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    password: CodecTypes['pg/text@1']['output'];
    updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    username: CodecTypes['pg/text@1']['output'] | null;
    tasks: public_Task[];
    readonly [RelationKeys]?: 'tasks';
  };
}